[Dev 1.2] Fixing size fixed on 5 bug
