+++ Game/StartGame.java
@@ -1,14 +1,14 @@
 package Game;
 
-import Game.Model.Controller;
+import Game.Model.GameController;
 import Game.View.View;
-import javafx.application.Application;
 import javafx.stage.Stage;
 
 
 public class StartGame {
+    View view;
     public StartGame(Stage stage, int size, int timer, boolean gameMode, int rounds){
-        Controller model = new Controller(size, timer, gameMode, rounds);
-        View view = new View(stage, model);
+        GameController model = new GameController(size, timer, gameMode, rounds);
+        view = new View(stage, model);
     }
 }
